// src/roles/entities/role.enum.ts
export enum Role {
  ADMIN = "admin",
  SITE_MANAGER = "site_manager",
  PETITION_MANAGEMENT_COMPANY = "petition_management_company",
  BALLOT_OR_INITIATIVE_COMMITTEE = "ballot_or_initiative_committee",
  POLITICAL_CANDIDATE = "political_candidate",
  PETITION_GATHERER = "petition_gatherer",
  PETITION_VALIDATOR = "petition_validator"
}